+++ PanoramicData.NCalcExtensions/Extensions/IsNullOrEmpty.cs
@@ -33,7 +33,7 @@ internal static class IsNullOrEmpty
 		}
 		catch (Exception e) when (e is not (NCalcExtensionsException or FormatException))
 		{
-			throw new FormatException(e.Message);
+			throw new FormatException(e.Message, e);
 		}
 	}
 }